Creating route and renaming files.
