+++ app/routes/toolkit_routes.py
@@ -5,8 +5,10 @@ directory of this distribution and at https://github.com/CivicActions/ssp-flask#
 
 from flask import redirect, request
 
-from app.routes import bp
+from app.routes.routes import bp
 from app.ssp_tools.createfiles import create_files
+
+# from app.ssp_tools.make_families import make_families
 
 
 @bp.route("/createfiles", defaults={"subpath": ""}, methods=["GET"])
@@ -14,4 +16,10 @@ from app.ssp_tools.createfiles import create_files
 def toolkit_render_files(subpath: str):
     to_render = subpath
     create_files(to_render=to_render)
+    return redirect(request.referrer or "/")
+
+
+@bp.route("/makefamilies", methods=["GET"])
+def toolkit_make_families():
+    # make_families()
     return redirect(request.referrer or "/")
